+++ src/Smartstore.Modules/Smartstore.WebApi/Controllers/Checkout/OrdersController.cs
@@ -75,6 +75,13 @@ namespace Smartstore.Web.Api.Controllers
             return GetRelatedEntity(key, x => x.ShippingAddress);
         }
 
+        [HttpGet("Orders({key})/WalletHistory"), ApiQueryable]
+        [Permission("Wallet.read")]
+        public IQueryable<WalletHistory> GetWalletHistory(int key)
+        {
+            return GetRelatedQuery(key, x => x.WalletHistory);
+        }
+
         [HttpGet("Orders({key})/DiscountUsageHistory"), ApiQueryable]
         [Permission(Permissions.Promotion.Discount.Read)]
         public IQueryable<DiscountUsageHistory> GetDiscountUsageHistory(int key)